requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
pyinstaller>=6.0.0
pytest-cov>=5.0